AMD F15: Fix warning in Proc/CPU/Feature
Fix Warning:
cpuFeatureLeveling.c:265, GNU Compiler 4 (gcc), Priority: Normal
cast to pointer from integer of different size [-Wint-to-pointer-cast]
with an intermediate cast to (intptr_t)

Reviewed-on: http://review.coreboot.org/3126
Tested-by: build bot (Jenkins)

+++ src/vendorcode/amd/agesa/f15/Proc/CPU/Feature/cpuFeatureLeveling.c
@@ -262,5 +262,5 @@ GetGlobalCpuFeatureListAddress (
 
   AddressValue = GLOBAL_CPU_FEATURE_LIST_TEMP_ADDR;
 
-  *Address = (UINT64 *)(AddressValue);
+  *Address = (UINT64 *) (intptr_t) (AddressValue);
 }
